Replace PooledClass with PooledObjectMixin
I'd originally intended for this to be a mixin rather than a concrete
part of any class hierarchy, plus the existing implementation had
issues when it came to subclasses (they'd get put into the same pool as
their parents).

The mixin approach is the same but correctly supports subclassing and
doesn't impose constraints on the inheritance chain.