Remove per-handler log message
This is incredibly spammy and as there's no log browser built
in it makes any use of the Logs tab practically impossible, as
when used by TRP we get events flying through this for practically
everything.

I'd recommend that any code that really needs to log handler should
instead do it themselves _within_ the handler functions, rather than
force it upon them all without a say in the matter.